move changes to home route
