+++ src/layouts/index.jsx
@@ -1,3 +1,2 @@
 export { default as MainLayout } from './MainLayout'
 export { default as LoginLayout } from './LoginLayout'
-export { default as Layout } from './Layout'
